+++ circe-yaml-common/src/main/scala/io/circe/yaml/common/Printer.scala
@@ -1,0 +1,33 @@
+package io.circe.yaml.common
+
+import io.circe.Json
+
+trait Printer {
+  def pretty(json: Json): String
+}
+
+object Printer {
+
+  sealed trait FlowStyle
+  object FlowStyle {
+    case object Flow extends FlowStyle
+    case object Block extends FlowStyle
+  }
+
+  sealed trait LineBreak
+  object LineBreak {
+    case object Unix extends LineBreak
+    case object Windows extends LineBreak
+    case object Mac extends LineBreak
+  }
+
+  sealed trait StringStyle
+  object StringStyle {
+    case object Plain extends StringStyle
+    case object DoubleQuoted extends StringStyle
+    case object SingleQuoted extends StringStyle
+    case object Literal extends StringStyle
+    case object Folded extends StringStyle
+  }
+
+}
