+++ circe-yaml-v12/src/main/scala/io/circe/yaml/v12/ParserImpl.scala
@@ -1,23 +1,18 @@
-package io.circe.yaml
+package io.circe.yaml.v12
 
 import cats.syntax.either._
 import io.circe._
+import io.circe.yaml.common
 import java.io.{ Reader, StringReader }
-import org.yaml.snakeyaml.LoaderOptions
-import org.yaml.snakeyaml.Yaml
-import org.yaml.snakeyaml.constructor.SafeConstructor
-import org.yaml.snakeyaml.nodes._
+import java.util.Optional
+import org.snakeyaml.engine.v2.api.LoadSettings
+import org.snakeyaml.engine.v2.composer.Composer
+import org.snakeyaml.engine.v2.constructor.StandardConstructor
+import org.snakeyaml.engine.v2.nodes._
+import org.snakeyaml.engine.v2.scanner.StreamReader
 import scala.collection.JavaConverters._
 
-final case class Parser(
-  maxAliasesForCollections: Int = 50
-) extends yaml.common.Parser {
-
-  private val loaderOptions = {
-    val options = new LoaderOptions()
-    options.setMaxAliasesForCollections(maxAliasesForCollections)
-    options
-  }
+class ParserImpl(settings: LoadSettings) extends common.Parser {
 
   /**
    * Parse YAML from the given [[Reader]], returning either [[ParsingFailure]] or [[Json]]
@@ -29,37 +24,50 @@ final case class Parser(
     json <- yamlToJson(parsed)
   } yield json
 
-  def parse(yaml: String): Either[ParsingFailure, Json] = parse(new StringReader(yaml))
+  def parse(yaml: String): Either[ParsingFailure, Json] =
+    parse(new StringReader(yaml))
 
   def parseDocuments(yaml: Reader): Stream[Either[ParsingFailure, Json]] = parseStream(yaml).map(yamlToJson)
   def parseDocuments(yaml: String): Stream[Either[ParsingFailure, Json]] = parseDocuments(new StringReader(yaml))
 
-  private[this] def parseSingle(reader: Reader) =
-    Either.catchNonFatal(new Yaml(loaderOptions).compose(reader)).leftMap(err => ParsingFailure(err.getMessage, err))
+  private[this] def asScala[T](ot: Optional[T]): Option[T] =
+    if (ot.isPresent) Some(ot.get()) else None
+
+  private[this] def createComposer(reader: Reader) =
+    new Composer(settings, new org.snakeyaml.engine.v2.parser.ParserImpl(settings, new StreamReader(settings, reader)))
+
+  private[this] def parseSingle(reader: Reader): Either[ParsingFailure, Node] =
+    Either.catchNonFatal {
+      val composer = createComposer(reader)
+      asScala(composer.getSingleNode)
+    } match {
+      case Left(err)          => Left(ParsingFailure(err.getMessage, err))
+      case Right(None)        => Left(ParsingFailure("no document found", new RuntimeException("no document found")))
+      case Right(Some(value)) => Right(value)
+    }
 
   private[this] def parseStream(reader: Reader) =
-    new Yaml(loaderOptions).composeAll(reader).asScala.toStream
+    createComposer(reader).asScala.toStream
 
   private[this] object CustomTag {
-    def unapply(tag: Tag): Option[String] = if (!tag.startsWith(Tag.PREFIX))
+    def unapply(tag: Tag): Option[String] = if (!tag.getValue.startsWith(Tag.PREFIX))
       Some(tag.getValue)
     else
       None
   }
 
-  private[this] class FlatteningConstructor extends SafeConstructor {
+  private[this] class FlatteningConstructor(settings: LoadSettings) extends StandardConstructor(settings) {
     def flatten(node: MappingNode): MappingNode = {
       flattenMapping(node)
       node
     }
 
-    def construct(node: ScalarNode): Object =
-      getConstructor(node).construct(node)
+    def construct(node: ScalarNode): Object = super.construct(node) // to make the method public
   }
 
   private[this] def yamlToJson(node: Node): Either[ParsingFailure, Json] = {
     // Isn't thread-safe internally, may hence not be shared
-    val flattener: FlatteningConstructor = new FlatteningConstructor
+    val flattener: FlatteningConstructor = new FlatteningConstructor(settings)
 
     def convertScalarNode(node: ScalarNode) = Either
       .catchNonFatal(node.getTag match {
@@ -83,7 +91,7 @@ final case class Parser(
         case Tag.NULL => Json.Null
         case CustomTag(other) =>
           Json.fromJsonObject(JsonObject.singleton(other.stripPrefix("!"), Json.fromString(node.getValue)))
-        case other => Json.fromString(node.getValue)
+        case _ => Json.fromString(node.getValue)
       })
       .leftMap { err =>
         ParsingFailure(err.getMessage, err)
@@ -126,8 +134,4 @@ final case class Parser(
       }
     }
   }
-}
-
-object Parser {
-  val default: Parser = Parser()
 }
